тесты: чистое ядро src/core и рабочее окружение native
Модель данных (Settings, AttinyData, CalculatedData, перечисления и
описывающие их константы) вынесена из setup.h и master_i2c.h в
src/core/types.h — заголовок без Arduino.h, собирается и в прошивку,
и в хостовые тесты.

Починено окружение native: test_filter = test_ota/* не совпадал с именем
сюита, из-за чего pio test молча пропускал все тесты и выходил с кодом 0.
Добавлены test_build_src и build_src_filter — без них ядро с .cpp не
попадёт в тестовый бинарник.

Тесты test_ota впервые реально выполняются: 11 тестов, все зелёные.
Поведение прошивки не меняется.
