+++ ESP8266/platformio.ini
@@ -146,11 +146,15 @@ platform_packages = platformio/tool-scons@~4.40801.0
 
 
 [env:native]
+# Хостовые тесты бизнес-логики: pio test -d ESP8266 -e native
+# Собирается только чистое ядро src/core (без Arduino), остальной src отсечён.
 platform = native
 test_framework = googletest
-build_flags = -std=c++17
+test_build_src = yes
+build_src_filter = -<*> +<core/*>
+test_filter = test_*
+build_flags = -std=c++17 -DFIRMWARE_VERSION=${this.firmware_version}
 lib_deps = ArduinoJson@7.3.1
-test_filter = test_ota/*
 platform_packages = platformio/tool-scons@~4.40801.0
 
 [secrets]
